_site/requirements-tools.txt
requests
bibtexparser
jinja2
pypdf
pyyaml
python-pptx
